Modificado dosificación, calificar, fechas de entrega.
-- Dosificación
Se agregaron todos los grupos en el modal de fechas de entrega.
Se agregó un valor por defecto de 0 en instrumentos y autocompletar el 100 si ya hay 2 instrumentos.
Se cambió la vista principal de dosificación a un DTO y se agregó el mensaje de 'sin realizar'
Se cambió la vista de edición de dosificación.
Se resolvió el error de colaboradores repetidos en el formato de impresión de dosificación compartida.
Se agregó un selector de profesor del cual se importa la dosificación.
Se agregó un límite de no edición en dosificaciones importadas.

-- Calificación
Se agregó un mensaje de información que muestra cuando no se está en el periodo actual.
Se agregaron clases de color en los alumnos con remediales y extraordinarios.

-- Reportes
Se solucionó el error de los parciales de los reportes de calificación, seguimiento y asistencias.

+++ model/dtoreport/CalificacionInstrumentoDTO.java
@@ -4,7 +4,7 @@ public interface CalificacionInstrumentoDTO {
 	
 	Integer getMecanismo();
 	
-	Integer getPonderacion();
+	Float getPonderacion();
 	
 	Integer getInstrumento();
